Handle the case where a property type is in fact an alias
This is seen for example in SSM::PatchBaseLine's PatchGroup.
See http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-ssm-patchbaseline.html VS the spec:
The former specifies a list of strings for the patch groups, whereas the
later specifies a list of PatchGroup.

In this case, we create an alias from this property type to an
expression of AWS primitive type